💄#406: Refactor header
