+++ apps/consent-ui/src/components/common/Header/HamburgerMenu/index.tsx
@@ -4,7 +4,7 @@ import styles from './HamburgerMenu.module.scss';
 
 const HamburgerMenu = ({ options }: { options: { label: React.ReactNode; link?: string }[] }) => {
 	return (
-		<div>
+		<div className={styles.hamburgerContainer}>
 			{options.map((option, index) => (
 				<Link href={option.link || ''} className={styles.hamburgerLine} key={index}>
 					{option.label}
